refactor(components): group standard components into a single folder
#28
