+++ src/components/DefaultComponents/DefaultLogo/tests/DefaultLogo.test.tsx
